requirements.txt
Flask==0.12.2
flask-sockets=0.2.1
websocket-client==0.47.0